+++ apps/api/src/app.module.ts
@@ -8,6 +8,7 @@ import { PostsModule } from './posts/posts.module';
 import { KolModule } from './kol/kol.module';
 import { TagsModule } from './tags/tags.module';
 import { UploadsModule } from './uploads/uploads.module';
+import { InteractionsModule } from './interactions/interactions.module';
 import { AllExceptionsFilter } from './common/filters/http-exception.filter';
 import { ResponseInterceptor } from './common/interceptors/response.interceptor';
 
@@ -24,6 +25,7 @@ import { ResponseInterceptor } from './common/interceptors/response.interceptor'
     KolModule,
     TagsModule,
     UploadsModule,
+    InteractionsModule,
   ],
   providers: [
     {